deepface_flask/requirements.txt
absl-py==1.2.0
alembic==0.8.10
align==0.0.11
aniso8601==9.0.1
annoy==1.17.1
ansible==6.5.0
ansible-core==2.13.5
apriori==1.0.0
asgiref==3.6.0
astor==0.8.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
Babel==2.11.0
BareNecessities==0.2.8
beautifulsoup4==4.11.1
bleach==1.4
blinker==1.3
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.4
ConfigArgParse==1.5.3
copyheaders==0.0.2
cryptography==38.0.1
cssselect==1.1.0
cycler==0.11.0
decorator==5.1.1
deepface==0.0.79
Deprecated==1.2.13
distlib==0.3.4
Django==4.2
django-bootstrap4==23.1
django-redis==5.2.0
djangorestframework==3.14.0
facenet==1.0.5
filelock==3.6.0
fire==0.5.0
Flask==2.2.2
Flask-Admin==1.5.0
Flask-BabelEx==0.9.4
Flask-Mail==0.9.1
Flask-MySQLdb==1.0.1
Flask-PageDown==0.1.4
Flask-RESTful==0.3.9
Flask-Script==2.0.5
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.2
flatbuffers==23.3.3
fonttools==4.37.1
gast==0.4.0
gdown==4.7.1
gensim==4.3.1
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.3.post0
grpcio==1.48.1
gunicorn==20.1.0
h5py==3.8.0
html5lib==1.0b3
idna==2.5
imageio==2.21.2
imageio-ffmpeg==0.4.7
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.16.0
keras==2.11.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==15.0.6.1
lxml==4.9.1
Mail==2.1.0
Mako==1.2.3
Markdown==2.6.11
MarkupSafe==2.1.1
matplotlib==3.5.3
mtcnn==0.1.1
mysqlclient==2.1.1
nltk==3.8.1
numpy==1.23.4
oauthlib==3.2.0
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
paddlepaddle==2.2.0
pandas==1.5.0
Pillow==9.4.0
platformdirs==2.5.1
protobuf==3.19.4
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pygame==2.1.0
pymongo==4.2.0
PyMySQL==1.0.2
pyparsing==3.0.9
pyquery==1.4.3
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
python-editor==1.0.3
pytz==2022.5
PyYAML==6.0
qiniu==7.2.0
rauth==0.7.2
redis==4.3.4
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
resolvelib==0.8.1
retina-face==0.0.13
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
smart-open==6.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
speaklater==1.3
SQLAlchemy==1.4.19
sqlparse==0.4.3
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.11.0
tensorflow-gpu==2.9.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
threadpoolctl==3.1.0
tk==0.1.0
tokenizers==0.13.2
torch @ file:///D:/%E8%BF%85%E9%9B%B7%E4%B8%8B%E8%BD%BD/torch-1.12.1%2Bcu116-cp39-cp39-win_amd64.whl
torchsummary==1.5.1
torchvision @ file:///D:/%E8%BF%85%E9%9B%B7%E4%B8%8B%E8%BD%BD/torchvision-0.13.1%2Bcu116-cp39-cp39-win_amd64.whl
tqdm==4.64.1
typing_extensions==4.3.0
tzdata==2023.3
urllib3==1.21.1
videofeatures-clipit==1.0
virtualenv==20.13.4
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==2.2.2
wget==3.2
wrapt==1.12.1
WTForms==2.1
xlwt==1.3.0
yapf==0.32.0
zipp==3.8.1
